requirements.txt
ipykernel
pdfplumber
python-docx 
pandas
nltk
gensim
pandas 
numpy
seaborn
matplotlib
tensorflow
scikit-learn==1.3.2
xgboost
streamlit==1.31.1
optuna 
joblib
